+++ src/index.ts
@@ -36,6 +36,7 @@ import {
   Session,
   SessionCreateParams,
   SessionCreateResponse,
+  SessionDebugParams,
   SessionListParams,
   SessionListResponse,
   SessionLiveURLs,
@@ -288,6 +289,7 @@ export declare namespace Browserbase {
     type SessionCreateParams as SessionCreateParams,
     type SessionUpdateParams as SessionUpdateParams,
     type SessionListParams as SessionListParams,
+    type SessionDebugParams as SessionDebugParams,
   };
 
   export {