+++ tests/api-resources/sessions/sessions.test.ts
@@ -64,7 +64,7 @@ describe('resource sessions', () => {
           ],
           proxySettings: { caCertificates: ['182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e'] },
           region: 'us-west-2',
-          timeout: 60,
+          api_timeout: 60,
           userMetadata: { foo: 'bar' },
         },
         { path: '/_stainless_unknown_path' },
@@ -149,5 +149,12 @@ describe('resource sessions', () => {
     await expect(client.sessions.debug('id', { path: '/_stainless_unknown_path' })).rejects.toThrow(
       Browserbase.NotFoundError,
     );
+  });
+
+  test('debug: request options and params are passed correctly', async () => {
+    // ensure the request options are being passed correctly by passing an invalid HTTP method in order to cause an error
+    await expect(
+      client.sessions.debug('id', { expiresIn: 60 }, { path: '/_stainless_unknown_path' }),
+    ).rejects.toThrow(Browserbase.NotFoundError);
   });
 });